hide literal node
it doesn't reaaally exist in the code, it doesn't need to
be an explicit node in the tree. This removes a bunch of nesting
in the tree and makes it simpler.

I know you've most likely already read this, but for the benefit of future me
and anyone else who might be interested,
http://tree-sitter.github.io/tree-sitter/creating-parsers#structuring-rules-well